requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
Django==5.1.5
django-countries==7.6.1
idna==3.10
narwhals==1.36.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
reportlab==4.3.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
